# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a documentation of SeahorseFi off-chain point tracker server.",
        "title": "SeahorseFi API Doc",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/wallets/points": {
            "get": {
                "description": "Get all wallets point",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Wallet"
                ],
                "summary": "Get all wallets point",
                "responses": {
                    "200": {
                        "description": "Successful response with list of wallets",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entities.Wallet"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "entities.Event": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "block_number": {
                    "type": "integer"
                },
                "closed_at": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "event_type": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "last_calculated_at": {
                    "type": "string"
                },
                "transaction_hash": {
                    "type": "string"
                },
                "wallet_id": {
                    "type": "string"
                }
            }
        },
        "entities.Wallet": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "event": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.Event"
                    }
                },
                "id": {
                    "type": "string"
                },
                "points": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        }
    }
}